Sample project restructuring
