+++ tests/testproject/spiders.py
@@ -1,5 +1,5 @@
 from zineb.app import Zineb
-from zineb.tests.testproject.models.base import SimpleModel
+from zineb.tests.testproject.models import SimpleModel
 
 # Create your spiders here
 
@@ -8,8 +8,7 @@ class MySpider(Zineb):
         'http://example.com'
     ]
 
-    def start(self, response, request, **kwargs):
+    def start(self, response, **kwargs):
         model = SimpleModel()
         model.add_value('url', response.find('a')['href'])
-        print(model)
